Backend/requirements.txt
Django==5.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-mongodb-backend==5.2.0b0
python-decouple==3.8
pymongo==4.7.2
